+++ cli/internal/ossiedir/osidir.go
@@ -1,4 +1,4 @@
-package osidir
+package ossiedir
 
 import (
 	"fmt"
