+++ cli/internal/ossiedir/osidir_test.go
@@ -1,4 +1,4 @@
-package osidir
+package ossiedir
 
 import (
 	"os"
